Update runsui.pug
Might be nice to have warmspot as an option for remote analysis shifters so we don't have tags that are "Hotspot" "Hot Spot" "Warm Spot" "Warmspot" and any combination that are all trying to do the same thing. I didn't check if I have to change anything other than add the option for this to work. I also thought it would be fun to try to do the pull request myself.